requirements.txt
asgiref==3.2.3
Django==3.0.3
pytz==2019.3
sqlparse==0.3.1
envparse==0.2.0
gunicorn==20.0.4
pytest==5.4.1
ipdb==0.13.2
pytest-django==3.9.0
WebTest==2.0.34
django-webtest==1.9.7
factory-boy==2.12.0
django-rest-framework==0.1.0
tqdm==4.48.2